Move list settings to a seperate file
